extract the logic as a function
